+++ src/main/java/blackjack/domain/gameplayer/Player.java
@@ -1,16 +1,21 @@
-package blackjack.domain;
+package blackjack.domain.gameplayer;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class Dealer implements Person {
-    private static final int DEALER_STOP_HIT_BOUND = 17;
+public class Player implements Person {
     private static final int BURST_NUMBER = 21;
+    private static final int ACE_MAX_NUMBER = 11;
     private static final int DIFFERENCE_WITH_ACE_NUMBER = 10;
+    private final Name name;
     private final List<Card> cards;
 
-    public Dealer() {
+    public Player(Name name) {
+        this.name = name;
         this.cards = new ArrayList<>();
     }
 
@@ -20,19 +25,28 @@ public class Dealer implements Person {
     }
 
     @Override
+    public List<Card> showCards() {
+        return Collections.unmodifiableList(cards);
+    }
+
+    public String showName() {
+        return name.getName();
+    }
+
+    @Override
     public boolean isHit() {
         int totalScore = calculateScore();
-        return totalScore < DEALER_STOP_HIT_BOUND;
+        return totalScore < BURST_NUMBER;
     }
 
     @Override
     public int calculateScore() {
         int totalScore = cards.stream()
-                .map(card -> Collections.max(card.getScore()))
+                .map(card -> Collections.min(card.getScore()))
                 .reduce(0, Integer::sum);
 
-        if (totalScore > BURST_NUMBER && hasACE()) {
-            return totalScore - DIFFERENCE_WITH_ACE_NUMBER;
+        if (totalScore <= ACE_MAX_NUMBER && hasACE()) {
+            return totalScore + DIFFERENCE_WITH_ACE_NUMBER;
         }
         return totalScore;
     }
@@ -40,14 +54,5 @@ public class Dealer implements Person {
     private boolean hasACE() {
         return cards.stream()
                 .anyMatch(card -> card.getCardNumberToString().equals(CardNumber.ACE.getNumber()));
-    }
-
-    @Override
-    public List<Card> showCards() {
-        return List.of(cards.get(0));
-    }
-
-    public List<Card> showAllCards() {
-        return Collections.unmodifiableList(cards);
     }
 }
