+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,5 +1,8 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSymbol;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
