+++ src/test/java/blackjack/domain/gameplayer/GamePlayerTest.java
@@ -1,9 +1,13 @@
-package blackjack.domain;
+package blackjack.domain.gameplayer;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSymbol;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
