+++ src/test/java/blackjack/domain/gameplayer/PlayerTest.java
@@ -1,9 +1,13 @@
-package blackjack.domain;
+package blackjack.domain.gameplayer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.List;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.CardSymbol;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
