+++ src/test/java/blackjack/domain/gameplayer/PlayersTest.java
@@ -1,10 +1,11 @@
-package blackjack.domain;
+package blackjack.domain.gameplayer;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import java.util.ArrayList;
 import java.util.List;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
